+++ lib/metro/models/metro_project_file.dart
@@ -1,0 +1,7 @@
+/// Model class used for creating files
+class MetroProjectFile {
+  String name;
+  String? creationPath;
+
+  MetroProjectFile(String name, {this.creationPath}) : this.name = name;
+}