+++ lib/widgets/ny_pull_to_refresh.dart
@@ -325,6 +325,7 @@ class _NyPullToRefreshState<T> extends NyState<NyPullToRefresh> {
     });
   }
 
+  /// Returns the header type
   Widget headerType() {
     switch (widget.headerStyle) {
       case "ClassicHeader":